Arena/MCP_servers/aider/requirements.txt
fastmcp==2.6.1
aider-chat
python-dotenv